pyproject.toml
[tool.poetry]
name = "dev-test"
version = "0.1.0"
description = ""
authors = ["Aradhya-Tripathi <aradhyatripathi51@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
requests = "^2.25.1"
python-dotenv = "^0.17.1"
Flask-SQLAlchemy = "^2.5.1"
autopep8 = "^1.5.7"
coverage = "^5.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
